[API] Union redesign
[API] removed rpos from all body constructors

Changes: Bodies that are part of a union now store all attributes relative to its respective coordinate frame. This affects gpos and q (Rotation), which are implicitly relative once a body becomes part of the union. These properties became private and the global Position can be accessed via getGlobalPos() which performs a recursive calcuation of the position of this body in the world frame. This avoids error-prone update procedures, if a Union is tranlated or rotated.