OpenMP support is now switchable within the generation script
Since MSVC only has beta support for OpenMP 2.0 firstprivate
cannot be used for these compilers. However, firstprivate
is necessary for a reasonably safe implementation of OpenMP
in MESA-PD. The solution is to move OpenMP support to the
generator and disable OpenMP for the checked in version.
The waLBerla OpenMP guards are removed as they are no longer
needed. It is now assumed that if MESA-PD is generated with
OpenMP support it is about to be used. If OpenMP is disabled
with CMake warnings will pop up about undefined pragmas.